Making router props available to TimeSlots component
In order to use components as is only TimeSlots component needs
to be changed a bit.

Only problem is, it requires information from routes which is not
available in the store since regular reducers don't get to use that.

So using store information from Provider's context and adding
router from props to VinicityReservationPage container's child context
allows this alterante rendering mode to function.

Mind you, in this commit said functionality is yet to be implemented.